🐚 UPDATE :: 댓글 반환 코드 수정

+++ src/main/java/com/dsm/up/domain/comment/service/CommentService.java
@@ -4,6 +4,9 @@ import com.dsm.up.domain.comment.domain.Comment;
 import com.dsm.up.domain.comment.domain.repository.CommentRepository;
 import com.dsm.up.domain.comment.exception.CommentNotFoundException;
 import com.dsm.up.domain.comment.presentation.dto.request.CommentRequest;
+import com.dsm.up.domain.user.exception.UserNotMatchException;
+import com.dsm.up.domain.user.service.util.UserUtil;
+import com.dsm.up.global.aws.S3Util;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -12,27 +15,30 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 public class CommentService {
     private final CommentRepository commentRepository;
+    private final UserUtil userUtil;
 
     @Transactional
     public Long creat(CommentRequest request) {
         return commentRepository.save(Comment.builder()
+                .user(userUtil.getUser())
                 .content(request.getContent())
-                //.user()
                 .build()).getId();
     }
 
-    @Transactional //comment 작성자와 일치하는지 확인
+    @Transactional
     public Long update(Long id, CommentRequest request) {
         Comment comment = commentRepository.findById(id)
                 .orElseThrow(() -> CommentNotFoundException.EXCEPTION);
+        if(!comment.getUser().getAccountId().equals(userUtil.getUserId())) throw UserNotMatchException.EXCEPTION;
 
         return comment.update(request.getContent());
     }
 
-    @Transactional //comment 작성자와 일치하는지 확인
+    @Transactional
     public void delete(Long id) {
         Comment comment = commentRepository.findById(id)
                 .orElseThrow(() -> CommentNotFoundException.EXCEPTION);
+        if(!comment.getUser().getAccountId().equals(userUtil.getUserId())) throw UserNotMatchException.EXCEPTION;
 
         commentRepository.delete(comment);
     }
